refactor: consolidate handler utilities
