+++ src/logic/operationHandlers/addComponentHandler.js
@@ -18,8 +18,8 @@ import {
   initHandlerLogger,
   validateDeps,
   getExecLogger,
-} from './handlerUtils.js';
-import { assertParamsObject } from '../../utils/handlerUtils.js';
+} from '../../utils/handlerUtils/service.js';
+import { assertParamsObject } from '../../utils/handlerUtils/params.js';
 
 /**
  * Parameters accepted by {@link AddComponentHandler#execute}.
