+++ src/logic/operationHandlers/addPerceptionLogEntryHandler.js
@@ -5,7 +5,7 @@
 
 import { PERCEPTION_LOG_COMPONENT_ID } from '../../constants/componentIds.js';
 import { DISPLAY_ERROR_ID } from '../../constants/eventIds.js';
-import { assertParamsObject } from '../../utils/handlerUtils.js';
+import { assertParamsObject } from '../../utils/handlerUtils/params.js';
 
 const DEFAULT_MAX_LOG_ENTRIES = 50;
