+++ src/logic/operationHandlers/dispatchPerceptibleEventHandler.js
@@ -10,7 +10,7 @@
 /** @typedef {import('../../interfaces/ISafeEventDispatcher.js').ISafeEventDispatcher} ISafeEventDispatcher */
 
 import { DISPLAY_ERROR_ID } from '../../constants/eventIds.js';
-import { assertParamsObject } from '../../utils/handlerUtils.js';
+import { assertParamsObject } from '../../utils/handlerUtils/params.js';
 
 const EVENT_ID = 'core:perceptible_event';
