+++ src/logic/operationHandlers/getNameHandler.js
@@ -17,7 +17,7 @@ import { DEFAULT_FALLBACK_CHARACTER_NAME } from '../../constants/textDefaults.js
 import { resolveEntityId } from '../../utils/entityRefUtils.js';
 import { setContextValue } from '../../utils/contextVariableUtils.js';
 import { DISPLAY_ERROR_ID } from '../../constants/eventIds.js';
-import { assertParamsObject } from '../../utils/handlerUtils.js';
+import { assertParamsObject } from '../../utils/handlerUtils/params.js';
 import { safeDispatchError } from '../../utils/safeDispatchError.js';
 import BaseOperationHandler from './baseOperationHandler.js';
