+++ src/logic/operationHandlers/hasComponentHandler.js
@@ -15,7 +15,7 @@ import { DISPLAY_ERROR_ID } from '../../constants/eventIds.js';
 import { resolveEntityId } from '../../utils/entityRefUtils.js';
 import { setContextValue } from '../../utils/contextVariableUtils.js';
 import storeResult from '../../utils/contextVariableUtils.js';
-import { assertParamsObject } from '../../utils/handlerUtils.js';
+import { assertParamsObject } from '../../utils/handlerUtils/params.js';
 
 /**
  * Parameters accepted by {@link HasComponentHandler#execute}.
