+++ src/logic/operationHandlers/ifCoLocatedHandler.js
@@ -13,7 +13,7 @@ import { POSITION_COMPONENT_ID } from '../../constants/componentIds.js';
 import { DISPLAY_ERROR_ID } from '../../constants/eventIds.js';
 import { resolveEntityId } from '../../utils/entityRefUtils.js';
 import { safeDispatchError } from '../../utils/safeDispatchError.js';
-import { assertParamsObject } from '../../utils/handlerUtils.js';
+import { assertParamsObject } from '../../utils/handlerUtils/params.js';
 
 class IfCoLocatedHandler {
   /** @type {ILogger} */
