+++ src/logic/operationHandlers/modifyComponentHandler.js
@@ -16,8 +16,8 @@ import {
   initHandlerLogger,
   validateDeps,
   getExecLogger,
-} from './handlerUtils.js';
-import { assertParamsObject } from '../../utils/handlerUtils.js';
+} from '../../utils/handlerUtils/service.js';
+import { assertParamsObject } from '../../utils/handlerUtils/params.js';
 
 /**
  * @typedef {object} EntityRefObject
