+++ src/utils/handlerUtils/index.js
@@ -1,0 +1,11 @@
+import { assertParamsObject } from './params.js';
+import { initHandlerLogger, validateDeps, getExecLogger } from './service.js';
+
+export { assertParamsObject, initHandlerLogger, validateDeps, getExecLogger };
+
+export default {
+  assertParamsObject,
+  initHandlerLogger,
+  validateDeps,
+  getExecLogger,
+};
